changed some precompiled header items
