+++ SpectatorViewPlugin/SpectatorViewPlugin/CalibrationBoardGenerator/CalibrationBoardGenerator.vcxproj.filters
@@ -15,14 +15,6 @@
     </Filter>
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="pch.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-  </ItemGroup>
-  <ItemGroup>
-    <ClCompile Include="pch.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
     <ClCompile Include="CalibrationBoardGenerator.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
