--- SpectatorViewPlugin/SpectatorViewPlugin/CalibrationBoardGenerator/pch.cpp
@@ -1,5 +1,0 @@
-// pch.cpp: source file corresponding to pre-compiled header; necessary for compilation to succeed
-
-#include "pch.h"
-
-// In general, ignore this file, but keep it around if you are using pre-compiled headers.
