+++ SpectatorViewPlugin/SpectatorViewPlugin/SharedFiles/pch.h
