+++ SpectatorViewPlugin/SpectatorViewPlugin/SpectatorViewPlugin/SpectatorViewPlugin.cpp
@@ -1,4 +1,6 @@
-﻿#include "SpectatorViewPlugin.h"
+﻿#include "..\SharedFiles\pch.h"
+
+#include "SpectatorViewPlugin.h"
 #include "..\SharedFiles\ArUcoMarkerDetector.h"
 #include "..\SharedFiles\Calibration.h"
 
@@ -105,7 +107,8 @@ extern "C" __declspec(dllexport) bool __stdcall InitializeCalibration()
     if (!calibration)
     {
         calibration = std::make_unique<Calibration>();
-        return calibration->Initialize();
+        calibration->Initialize();
+        return true;
     }
 
     return true;
@@ -116,7 +119,8 @@ extern "C" __declspec(dllexport) bool __stdcall ResetCalibration()
 {
     if (calibration)
     {
-        return calibration->Initialize();
+        calibration->Initialize();
+        return true;
     }
 
     return false;
