+++ SpectatorViewPlugin/SpectatorViewPlugin/SpectatorViewPlugin/SpectatorViewPlugin.vcxproj.filters
@@ -27,9 +27,9 @@
     <ClInclude Include="..\SharedFiles\Calibration.h">
       <Filter>Shared Files</Filter>
     </ClInclude>
-    <ClInclude Include="..\SharedFiles\SharedInclude.h">
+    <ClInclude Include="targetver.h" />
+    <ClInclude Include="..\SharedFiles\pch.h">
       <Filter>Shared Files</Filter>
     </ClInclude>
-    <ClInclude Include="targetver.h" />
   </ItemGroup>
 </Project>
