+++ SpectatorViewPlugin/SpectatorViewPlugin/SpectatorViewPlugin/dllmain.cpp
@@ -1,4 +1,4 @@
-﻿#include "..\SharedFiles\SharedInclude.h"
+﻿#include "..\SharedFiles\pch.h"
 
 BOOL APIENTRY DllMain(HMODULE /* hModule */, DWORD ul_reason_for_call, LPVOID /* lpReserved */)
 {
